sysinfo::common

Module system

Macros

pid_decl 🔒

Structs

CGroupLimits

Contains memory limits for the current process.

Cpu

Contains all the methods of the Cpu struct.

CpuRefreshKind

Used to determine what you want to refresh specifically on the Cpu type.

LoadAvg

A struct representing system load average value.

MemoryRefreshKind

Used to determine which memory you want to refresh specifically.

Motherboard

This type allows to retrieve motherboard-related information.

Pid

Process ID.

Process

Struct containing information of a process.

ProcessRefreshKind

Used to determine what you want to refresh specifically on the Process type.

Product

This type allows to retrieve product-related information.

RefreshKind

Used to determine what you want to refresh specifically on the System type.

System

Type containing system’s information such as processes, memory and CPU.

Enums

KillError

Enum describing possible Process::kill_and_wait errors.

ProcessStatus

Enum describing the different status of a process.

ProcessesToUpdate

This enum allows you to specify if you want all processes to be updated or just some of them.

Signal

An enum representing signals on UNIX-like systems.

ThreadKind

Enum describing the different kind of threads.

UpdateKind

This enum allows you to specify when you want the related information to be updated.

Functions

get_current_pid

Returns the pid for the current process.