Enum sysinfo::Signal

pub enum Signal {

    Hangup,
    Interrupt,
    Quit,
    Illegal,
    Trap,
    Abort,
    IOT,
    Bus,
    FloatingPointException,
    Kill,
    User1,
    Segv,
    User2,
    Pipe,
    Alarm,
    Term,
    Child,
    Continue,
    Stop,
    TSTP,
    TTIN,
    TTOU,
    Urgent,
    XCPU,
    XFSZ,
    VirtualAlarm,
    Profiling,
    Winch,
    IO,
    Poll,
    Power,
    Sys,

}

An enum representing signals on UNIX-like systems.

On non-unix systems, this enum is mostly useless and is only there to keep coherency between the different OSes.

If you want the list of the supported signals on the current system, use SystemExt::SUPPORTED_SIGNALS.

Variants

Hangup

Hangup detected on controlling terminal or death of controlling process.

Interrupt

Interrupt from keyboard.

Quit

Quit from keyboard.

Illegal

Illegal instruction.

Trap

Trace/breakpoint trap.

Abort

Abort signal from C abort function.

IOT

IOT trap. A synonym for SIGABRT.

Bus

Bus error (bad memory access).

FloatingPointException

Floating point exception.

Kill

Kill signal.

User1

User-defined signal 1.

Segv

Invalid memory reference.

User2

User-defined signal 2.

Pipe

Broken pipe: write to pipe with no readers.

Alarm

Timer signal from C alarm function.

Term

Termination signal.

Child

Child stopped or terminated.

Continue

Continue if stopped.

Stop

Stop process.

TSTP

Stop typed at terminal.

TTIN

Terminal input for background process.

TTOU

Terminal output for background process.

Urgent

Urgent condition on socket.

XCPU

CPU time limit exceeded.

XFSZ

File size limit exceeded.

VirtualAlarm

Virtual alarm clock.

Profiling

Profiling time expired.

Winch

Windows resize signal.

IO

I/O now possible.

Poll

Pollable event (Sys V). Synonym for IO

Power

Power failure (System V).

Doesn’t exist on apple systems so will be ignored.

Sys

Bad argument to routine (SVr4).