Trait sysinfo::ProcessExt

pub trait ProcessExt: Debug {

    fn kill(&self) -> bool;

    fn kill_with(&self, signal: Signal) -> Option<bool>;

    fn name(&self) -> &str;

    fn cmd(&self) -> &[String];

    fn exe(&self) -> &Path;

    fn pid(&self) -> Pid;

    fn environ(&self) -> &[String];

    fn cwd(&self) -> &Path;

    fn root(&self) -> &Path;

    fn memory(&self) -> u64;

    fn virtual_memory(&self) -> u64;

    fn parent(&self) -> Option<Pid>;

    fn status(&self) -> ProcessStatus;

    fn start_time(&self) -> u64;

    fn run_time(&self) -> u64;

    fn cpu_usage(&self) -> f32;

    fn disk_usage(&self) -> DiskUsage;

}

Contains all the methods of the Process struct.

Required methods

fn kill(&self) -> bool

Sends Signal::Kill to the process (which is the only signal supported on all platforms by this crate).

If you want to send another signal, take a look at ProcessExt::kill_with.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    process.kill();
}

fn kill_with(&self, signal: Signal) -> Option<bool>

Sends the given signal to the process. If the signal doesn’t exist on this platform, it’ll do nothing and will return None. Otherwise it’ll return if the signal was sent successfully.

If you just want to kill the process, use ProcessExt::kill directly.

use sysinfo::{ProcessExt, Signal, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    if process.kill_with(Signal::Kill).is_none() {
        eprintln!("This signal isn't supported on this platform");
    }
}

fn name(&self) -> &str

Returns the name of the process.

⚠️ Important ⚠️

On linux, there are two things to know about processes’ name:

It is limited to 15 characters.

It is not always the exe name.

If you are looking for a specific process, unless you know what you are doing, in most cases it’s better to use ProcessExt::exe instead (which can be empty sometimes!).

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{}", process.name());
}

fn cmd(&self) -> &[String]

Returns the command line.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{:?}", process.cmd());
}

fn exe(&self) -> &Path

Returns the path to the process.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{}", process.exe().display());
}

fn pid(&self) -> Pid

Returns the pid of the process.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{}", process.pid());
}

fn environ(&self) -> &[String]

Returns the environment variables of the process.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{:?}", process.environ());
}

fn cwd(&self) -> &Path

Returns the current working directory.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{}", process.cwd().display());
}

fn root(&self) -> &Path

Returns the path of the root directory.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{}", process.root().display());
}

fn memory(&self) -> u64

Returns the memory usage (in KB).

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{} KB", process.memory());
}

fn virtual_memory(&self) -> u64

Returns the virtual memory usage (in KB).

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{} KB", process.virtual_memory());
}

fn parent(&self) -> Option<Pid>

Returns the parent pid.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{:?}", process.parent());
}

fn status(&self) -> ProcessStatus

Returns the status of the processus.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{:?}", process.status());
}

fn start_time(&self) -> u64

Returns the time where the process was started (in seconds) from epoch.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("Started at {} seconds", process.start_time());
}

fn run_time(&self) -> u64

Returns for how much time the process has been running (in seconds).

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("Running since {} seconds", process.run_time());
}

fn cpu_usage(&self) -> f32

Returns the total CPU usage (in %). Notice that it might be bigger than 100 if run on a multicore machine.

If you want a value between 0% and 100%, divide the returned value by the number of CPU processors.

Warning: If you want accurate CPU usage number, better leave a bit of time between two calls of this method (200 ms for example).

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    println!("{}%", process.cpu_usage());
}

fn disk_usage(&self) -> DiskUsage

Returns number of bytes read and written to disk.

⚠️ On Windows and FreeBSD, this method actually returns ALL I/O read and written bytes.

use sysinfo::{ProcessExt, System, SystemExt};

let s = System::new();
if let Some(process) = s.process(1337) {
    let disk_usage = process.disk_usage();
    println!("read bytes   : new/total => {}/{}",
        disk_usage.read_bytes,
        disk_usage.total_read_bytes,
    );
    println!("written bytes: new/total => {}/{}",
        disk_usage.written_bytes,
        disk_usage.total_written_bytes,
    );
}