sysinfo::unix::linux::component

Struct ComponentInner

pub(crate) struct ComponentInner {
    device_model: Option<String>,
    id: Option<String>,
    name: String,
    temperature: Option<f32>,
    max: Option<f32>,
    threshold_critical: Option<f32>,
    sensor_type: Option<ThermalSensorType>,
    label: String,
    input_file: Option<PathBuf>,
    highest_file: Option<PathBuf>,
    pub(crate) updated: bool,
}

Fields
device_model: Option<String>

Optional associated device of a Component.

id: Option<String>

ID of a Component.

name: String

The chip name.

Kernel documentation extract:

This should be a short, lowercase string, not containing
whitespace, dashes, or the wildcard character '*'.
This attribute represents the chip name. It is the only
mandatory attribute.
I2C devices get this attribute created automatically.

temperature: Option<f32>

Temperature current value

Read in: temp[1-*]_input.

Unit: read as millidegree Celsius converted to Celsius.

max: Option<f32>

Maximum value computed by sysinfo.

threshold_critical: Option<f32>

Critical threshold provided by the chip/kernel previous user write. Read in temp[1-*]_crit: Typically greater than corresponding temp_max values.

Unit: read as millidegree Celsius converted to Celsius.

sensor_type: Option<ThermalSensorType>

Sensor type, not common but can exist!

Read in: temp[1-*]_type Sensor type selection. Values integer:

1: CPU embedded diode

2: 3904 transistor

3: thermal diode

4: thermistor

5: AMD AMDSI

6: Intel PECI

Not all types are supported by all chips.

label: String

Component Label

Linux implementation details

read n: temp[1-*]_label Suggested temperature channel label. Value: Text string

Should only be created if the driver has hints about what this temperature channel is being used for, and user-space doesn’t. In all other cases, the label is provided by user-space.

input_file: Option<PathBuf>

File to read current temperature shall be temp[1-*]_input It may be absent but we don’t continue if absent.

highest_file: Option<PathBuf>

temp[1-*]_highest file to read if available highest value.

updated: bool
Implementations

impl ComponentInner

fn update_from(&mut self, _: Component)

impl ComponentInner

fn from_hwmon(components: &mut Vec<Component>, folder: &Path) -> Option<()>

Read out hwmon info (hardware monitor) from folder to get values’ path to be used on refresh as well as files containing max, critical value and label. Then we store everything into components.

Note that a thermal Component must have a way to read its temperature. If not, it will be ignored and not added into components.

What is read:

Mandatory: name the name of the hwmon.

Mandatory: tempN_input Drop Component if missing

Optional: sensor label, in the general case content of tempN_label see below for special cases

Optional: label

Optional: /device/model

Optional: highest historic value in tempN_highest.

Optional: max threshold value defined in tempN_max

Optional: critical threshold value defined in tempN_crit

Where N is a u32 associated to a sensor like temp1_max, temp1_input.

Doc to Linux kernel API.

Kernel hwmon API: https://www.kernel.org/doc/html/latest/hwmon/hwmon-kernel-api.html DriveTemp kernel API: https://docs.kernel.org/gpu/amdgpu/thermal.html#hwmon-interfaces Amdgpu hwmon interface: https://www.kernel.org/doc/html/latest/hwmon/drivetemp.html

fn format_label(&self, class: &str, id: u32) -> String

Compute a label out of available information. See the table in Component::label’s documentation.

pub(crate) fn temperature(&self) -> Option<f32>

pub(crate) fn max(&self) -> Option<f32>

pub(crate) fn critical(&self) -> Option<f32>

pub(crate) fn label(&self) -> &str

pub(crate) fn id(&self) -> Option<&str>

pub(crate) fn refresh(&mut self)