Crate sysinfo

For crate-level documentation, all features need to be enabled.

Modules

common 🔒

debug 🔒

macros 🔒

unix 🔒

utils 🔒

Structs

Disk

Struct containing a disk information.

DiskRefreshKind

Used to determine what you want to refresh specifically on the Disk type.

DiskUsage

Type containing read and written bytes.

Disks

Disks interface.

Enums

DiskKind

Enum containing the different supported kinds of disks.

Constants

IS_SUPPORTED_SYSTEM

Returns true if this OS is supported. Please refer to the crate-level documentation to get the list of supported OSes.

Functions

set_open_files_limit

This function is only used on Linux targets, when the system feature is enabled. In other cases, it does nothing and returns false.